+++ src/main/java/kr/happytravel/erp/sales/dao/AgencyDao.java
@@ -1,6 +1,6 @@
 package kr.happytravel.erp.sales.dao;
 
-import kr.happytravel.erp.sales.model.AgencyModel;
+import kr.happytravel.erp.sales.model.sales.AgencyModel;
 import java.util.List;
 import java.util.Map;
