+++ src/main/java/kr/happytravel/erp/sales/dao/FlightDao.java
@@ -2,7 +2,7 @@ package kr.happytravel.erp.sales.dao;
 
 
 
-import kr.happytravel.erp.sales.model.FlightModel;
+import kr.happytravel.erp.sales.model.sales.FlightModel;
 
 import java.util.List;
 import java.util.Map;
