+++ src/main/java/kr/happytravel/erp/sales/model/sales/FlightModel.java
@@ -1,4 +1,4 @@
-package kr.happytravel.erp.sales.model;
+package kr.happytravel.erp.sales.model.sales;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
@@ -11,21 +11,24 @@ import lombok.NoArgsConstructor;
 
 import java.util.Date;
 
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @JsonNaming(PropertyNamingStrategy.SnakeCaseStrategy.class)
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class SaleHistoryModel {
-    private int seq;
-    private String packageCode;
-    // 인원수
-    private int count;
-    private String paymentCode;
+public class FlightModel {
+
+    private String flightCode;
+    private String airline;
+    private String flightNumber;
+    private String phone;
+    private String departure;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyyMMddHHmm")
-    private Date paymentDate;
-
-
+    private Date departureTime;
+    private String destination;
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyyMMddHHmm")
+    private Date arrivalTime;
+    private int price;
+    private String isUsed;
 }
