+++ src/main/java/kr/happytravel/erp/sales/service/HotelService.java
@@ -1,6 +1,6 @@
 package kr.happytravel.erp.sales.service;
 
-import kr.happytravel.erp.sales.model.HotelModel;
+import kr.happytravel.erp.sales.model.sales.HotelModel;
 import java.util.List;
 import java.util.Map;
