+++ src/main/resources/sql/sales/HotelMapper.xml
@@ -5,7 +5,7 @@
 <mapper namespace="kr.happytravel.erp.sales.dao.HotelDao">
 
     <!-- 전체 호텔 목록 조회 -->
-    <select id="getHotelList" resultType="kr.happytravel.erp.sales.model.HotelModel" parameterType="hashmap">
+    <select id="getHotelList" resultType="kr.happytravel.erp.sales.model.sales.HotelModel" parameterType="hashmap">
         SELECT
         hotel_code,
         hotel_name,
@@ -29,7 +29,7 @@
     </select>
 
     <!-- 단건조회 -->
-    <select id="selectHotel" resultType="kr.happytravel.erp.sales.model.HotelModel">
+    <select id="selectHotel" resultType="kr.happytravel.erp.sales.model.sales.HotelModel" parameterType="hashmap">
         SELECT
         hotel_code,
         hotel_name,
@@ -44,7 +44,7 @@
     </select>
 
     <!-- 신규등록 -->
-    <insert id="insertHotel">
+    <insert id="insertHotel" parameterType="hashmap">
         INSERT INTO TB_HOTEL_INFO (
         hotel_code,
         hotel_name,
@@ -67,7 +67,7 @@
     </insert>
 
     <!-- 단건수정 -->
-    <update id="updateHotel">
+    <update id="updateHotel" parameterType="hashmap">
         UPDATE TB_HOTEL_INFO
         SET
         hotel_name = #{hotel_name},
@@ -82,7 +82,7 @@
     </update>
 
     <!-- 단건삭제 -->
-    <delete id="deleteHotel">
+    <delete id="deleteHotel" parameterType="hashmap">
         DELETE FROM TB_HOTEL_INFO WHERE hotel_code = #{hotel_code}
     </delete>
